[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.16 ms, Memory: 3.94 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 vector<long long> solution(int x, int n)
 {
-    int _x = x;
+    long long _x = x;
     vector<long long> answer;
     for (int i = 0; i < n; i++)
     {
